Moved test interfaces back to their tests themselves
Since IBM JDK 6 is not throwing errors about these anymore, might as
well move them back to the tests they belong to.

+review REVIEW-5658
