+++ subprojects/platform-base/src/test/groovy/org/gradle/platform/base/internal/registry/ComponentBinariesModelRuleExtractorTest.groovy
@@ -15,10 +15,7 @@
  */
 
 package org.gradle.platform.base.internal.registry
-import org.gradle.language.base.internal.testinterfaces.RawLibrary
-import org.gradle.language.base.internal.testinterfaces.SomeBinarySpec
-import org.gradle.language.base.internal.testinterfaces.SomeBinarySubType
-import org.gradle.language.base.internal.testinterfaces.SomeLibrary
+
 import org.gradle.language.base.plugins.ComponentModelBasePlugin
 import org.gradle.model.InvalidModelRuleDeclarationException
 import org.gradle.model.ModelMap
@@ -92,6 +89,14 @@ class ComponentBinariesModelRuleExtractorTest extends AbstractAnnotationModelRul
         "returnValue"             | "Method annotated with @ComponentBinaries must not have a return value."                                                                      | "non void method"
         "rawModelMap"             | "Parameter of type '${ModelMap.simpleName}' must declare a type parameter extending 'BinarySpec'."                                            | "non typed ModelMap parameter"
     }
+
+    static interface SomeBinarySpec extends BinarySpec {}
+
+    static interface SomeLibrary extends ComponentSpec {}
+
+    static interface RawLibrary extends ComponentSpec {}
+
+    static interface SomeBinarySubType extends SomeBinarySpec {}
 
     static class Rules {
         @ComponentBinaries
